# week3/api-automation/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: ShopEasy Order Management API
  description: |
    **QA Assessment API** — ShopEasy E-Commerce Platform
    
    This API covers the full order lifecycle: Authentication → Products → Cart → Orders → Payments.
    
    **Getting Started:**
    1. Call `POST /auth/login` with `admin@shopeasy.com` / `password123`
    2. Copy the `token` from the response
    3. Click **Authorize** (top right) and enter `Bearer <your-token>`
    4. You can now access all protected endpoints
  version: "1.0.0"
  contact:
    name: QA Assessment Support

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Auth
    description: User registration and login
  - name: Products
    description: Browse product catalog
  - name: Cart
    description: Manage shopping cart
  - name: Orders
    description: Place and manage orders
  - name: Payments
    description: Process and check payments

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: admin@shopeasy.com
        password:
          type: string
          example: password123

    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          example: john@example.com
        password:
          type: string
          example: securepass123
        name:
          type: string
          example: John Doe

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Wireless Headphones
        price:
          type: number
          example: 99.99
        category:
          type: string
          example: electronics
        stock:
          type: integer
          example: 50

    CartItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2

    Order:
      type: object
      properties:
        orderId:
          type: string
          example: ORD-1001
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]
        total:
          type: number
          example: 199.98
        items:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time

    PaymentRequest:
      type: object
      required: [orderId, method]
      properties:
        orderId:
          type: string
          example: ORD-1001
        method:
          type: string
          enum: [credit_card, debit_card, upi, net_banking]
          example: credit_card

    Error:
      type: object
      properties:
        error:
          type: string
          example: Resource not found

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and get Bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: tok_abc123xyz
                  userId:
                    type: integer
                    example: 1
                  message:
                    type: string
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  message:
                    type: string
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /products:
    get:
      tags: [Products]
      summary: List all products with optional filters
      parameters:
        - in: query
          name: category
          schema:
            type: string
          example: electronics
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cartTotal:
                    type: number
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart:
    get:
      tags: [Cart]
      summary: View current cart
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  subtotal:
                    type: number
                  itemCount:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart/items/{itemId}:
    delete:
      tags: [Cart]
      summary: Remove item from cart
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
          description: productId of the item to remove
          example: 1
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cartTotal:
                    type: number
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Item not in cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders:
    post:
      tags: [Orders]
      summary: Place order from current cart
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                  total:
                    type: number
                  status:
                    type: string
                  message:
                    type: string
        "400":
          description: Cart is empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
          example: ORD-1001
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden — not your order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders/{orderId}/cancel:
    delete:
      tags: [Orders]
      summary: Cancel an order
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
          example: ORD-1001
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "422":
          description: Cannot cancel shipped/delivered order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /payments:
    post:
      tags: [Payments]
      summary: Initiate payment for an order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        "201":
          description: Payment successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId:
                    type: string
                  status:
                    type: string
                  amount:
                    type: number
                  message:
                    type: string
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Payment already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Get payment status by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
          example: PAY-5001
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId:
                    type: string
                  orderId:
                    type: string
                  method:
                    type: string
                  amount:
                    type: number
                  status:
                    type: string
                  processedAt:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'